refactor promo api
