+++ app/api/promo/reorder/route.ts
@@ -1,6 +1,6 @@
 // ✅ Путь: app/api/promo/reorder/route.ts
 import { NextRequest, NextResponse } from 'next/server';
-import { supabaseAdmin } from '@/lib/supabase/server';
+import { prisma } from '@/lib/prisma';
 
 export async function POST(req: NextRequest) {
   try {
@@ -10,19 +10,11 @@ export async function POST(req: NextRequest) {
       return NextResponse.json({ error: 'Некорректный формат данных' }, { status: 400 });
     }
 
-    // Обновляем order_index в Supabase
-    const updates = order.map(({ id, order_index }: { id: number; order_index: number }) =>
-      supabaseAdmin
-        .from('promo_blocks')
-        .update({ order_index })
-        .eq('id', id)
+    await prisma.$transaction(
+      order.map(({ id, order_index }: { id: number; order_index: number }) =>
+        prisma.promo_blocks.update({ where: { id: Number(id) }, data: { order_index } })
+      )
     );
-
-    const results = await Promise.all(updates);
-    const errors = results.filter(result => result.error);
-    if (errors.length > 0) {
-      throw new Error(errors[0].error?.message || 'Ошибка обновления порядка');
-    }
 
     return NextResponse.json({ success: true });
   } catch (error: any) {
