+++ app/api/promo/validate/route.ts
